ipc/client: Refactor parallel client calls to fix close bugs.
We must ensure all flows that we've connected other than the one
we've chosen for the flowClient are closed.  This is somewhat
confusing because there are multiple codepaths.

The strategy is to put the cleanup logic in a separate function,
that is always invoked in a separate goroutine.  This makes it
easier to ensure that regardless of the codepath, cleanup always
occurs.

Change-Id: Ieb40d3a33532d7caddcd417f721d21a209438ab9