veyron2/vdl,ipc: Add new Signature support (incomplete).
The vdl side of this change removes the client and server stub
GetMethodTags, and adds a server stub Describe__.  The concept is
that the any object may implement Describe__ to describe the
interfaces it implements; stuff like interface names, tags,
documentation, variable names, etc; everything we can't get via
reflection.

The ReflectInvoker now uses Describe__ to grab all of this
information, and combines it with its reflection information
(which provides the exact argument types) to implement Signature
and MethodSignature.  One main principle is that the
ReflectInvoker will always return the exact type information it
extracts via reflection; the Describe__ information is only used
to enhance the type information, but never replaces it.

The ipc server has been extended to have general support for
reserved methods.  This is largely based on the glob support, but
just generalized.  On this side I also performed some refactoring
to help me understand the code.  I also added mutableCall and
mutableContext, which help make it more obvious what state we're
mutating; it gets confusing when we have recursive calls into the
same underlying flowServer.

The reserved __Signature and __MethodSignature methods don't
actually work yet; we need vom2 support for that to happen.

Change-Id: I23f2f1795d0383845ba53f2a2cb672c7b00e0d50