"veyron/security": Make principal serializaiton backwards compatible
CL 6820 changed the serialization format for BlessingStores, as a result
of which principals serialized in the old format could no longer be
deserialized.

This turned out to be huge inconvenience as it required re-generating
credentials for all envyor binaries (pis, identity server, etc).

This CL restores support for the old serialization format. In
particular,
* Principals serialized according to the old format can be deserialized
* However, subsequent updates to the principal's state get saved in the
new format.

Once everyone has switched to the new serialization format, we will
remove support for the old one.

Change-Id: Icd1204c27e93fbb3cd710b072dbc7b621a47f2c8