veyron/services/mgmt/node/impl: fixing source of test flakiness: insufficient timeout
Long-running subprocesses started by the node manager test (like the node
manager or mounttable subprocesses) occasionally exceed their default 1m
timeout and are terminated. This causes the test to fail, e.g. with the
"failed: ipc: response header decoding failed: EOF" message seen occasionally
when running the test (go/vissues/395).

This cl ups the timeout to 2 mins.

While we're at it, also fix cleanupDir in test mode to skip attempting to remove
the dir if it's "" (unrelated to the above, but annoying to see the warnings
in the test log).

Change-Id: I0b3185d03e323b2fa28cda6eaffcb735b157a09e