"veyron/security": BlessingStore bug
CL 7030 changed the blessingstore so that it saves all its state
right after the first read. This was to ensure that all state serialized
in the old format gets automatically updated to the new format.

However, this results in new blessingstore.sig files in our testdata
("veyron/security/testdata") on every test run.

This CL fixes the issue by doing the save only if we detect that the
existing serialized data is in the old format.

Change-Id: Iafd4e5309cd5175fcc16c09223488b85a0fb4a70