veyron/services/mgmt/node/impl: double timeout values.
We use timeout values to avoid waiting forever for children to start or to
finish. These should be generous enough for most cases, but occasionally I saw
test failures where it looked like a timeout fired too soon. This change takes
our quasi-arbitrary timeouts and doubles them. We should eventually have a more
principled way to set the timeouts. For now, we consolidate the various timeout
values and use constants to define and use them consistently. A related change:
remove the Wait preceding the Clean in the error recovery path of
testNodeManager (its utility was questionable, given that we already had a long
enough wait with WaitForReady).

Change-Id: Ic87a129f924b8fb837c6be1f0f6686971285164f