veyron2/ipc/model.go: Change VChildrenGlobber interface
This change makes the VChildrenGlobber interface incremental. The
implementations must now return a channel instead of a slice of strings,
which lets them to send the list of children incrementally.

The new interface is:
type VChildrenGlobber interface {
  GlobChildren__() (<-chan string, error)
}

As a side-effect, the framework's Glob logic must now perform a breadth-first
search to allow the channels to completely drain before moving on to other
objects.

Change-Id: Iabbf0a6ec7e27f34d86cbcb41b565fc628284dc9