veyron2/ipc/model.go: Normalize the Glob interfaces
With this change, the Glob__ and GlobChildren__ methods both take a
ServerContext as first argument and return a channel.

All the Glob() implementations are updated to use the AllGlobber
interface from ipc/model.go. This means that the mounttable.Globbable
interface is no longer needed and can now be removed.

The VGlob interface is renamed to Globber.

Change-Id: Idf1cd4663375eb3f1fc3baf7ca7606f59f7ded19