veyron.io/veyron/veyron/tools/principal: Introduce "principal store addtoroots" subcommand.
This CL introduces a "store addtoroots" subcommand to the "principal" command.

Expected use to allow credentials directory A to understand a blessing from
credentials directory B:
	principal -veyron.credentials B bless A arbitrary_label_that_is__ignored |
	principal -veyron.credentials A store addtoroots -

Several other subcommands call AddToRoots(), but always as a (sometimes
optional) side-effect of some other action.  This command does nothing bu the
AddToRoots(), making it possible for two unrelated principals to use the name
of the other in ACLs, without blessing one another.

The implementation is very similar to the "store set" subcommand, which is
immediately above it in the file.  The differences are:
- the help text,
- the new subcomment does not all Set(), and
- the call to AddToRoots() is not conditional on the -add_to_roots flag.

Change-Id: I2d23454815dcc319927ebe3912fc3007a7db412f