namespace: ResolveX smarter about errors.
We now only avoid returning an error when the destination error
is for ResolveX method not implemented.   We wouldn't even
need theat except for the __debug part of the space on a mounttable.
In that case the EP says the server is a mounttable but the __debug
stuff is intercepted by something that doesn't answer ResolveX.
Given that this extra work only happens for debugging, its no big
deal that we do the extra step.

Change-Id: I416a1a5f81c6f29bf2eaab7421acf7438feebc81