veyron/services/mgmt/node/impl: fix globbing on installations
The issue is that globbing on apps/<app title>/* or apps/... fails to reveal
installations of the app that do not have any instances.

The fix is to include a node for the installation itself in the tree of nodes
created in GlobChildren__, such that apps/<app title>/<installation id> will
still be found even if <installation id> has an empty set of children.

To cover this case, adding an example to the unit test of an installation
without instances, and showing that it correctly shows up in Globbing.

Change-Id: Id4b201059521059ce4cac5fce17d0c3f66fbffeb