veyron/services/mgmt/node/impl: use exec in the node manager wrapper script
The script wrapping the node manager invocation has no other purpose than to
start the node manager with the right env and args. Using exec ensures that we
won't have a parent process (the shell script) in the way. This helps ensure
that the node manager process gets the signals sent to it, e.g. by the agent
(which is now the case following http://go/vcl/7680).

Change-Id: I7d7e9f3e8006cfd42f0cd94a95eab544626d8d9d