v/services/mgmt/node/noded,v/tools/mgmt/nminstall: args for node manager
This change provides the node manager will optional arguments when using the
self-install feature: all unparsed args will be passed into the script generated
for the node manager by the installer.  This generic mechanism replaces the
special-casing we had for the publish name flag.

nminstall is updated to pass through optional args to the node manager
self-installer.

While at it, also fix a bug (in how we handled the SINGLE_USER variable).

Also, change nminstall to use exec when running the node manager at the end (as
per Robin's suggestion).

Change-Id: Ib7962774c61e90311a2fd9bc93ef3ee19b2b5ae9