"veyron/lib/testutil/security": ForkCredentials
In order to make it easy for tests to setup credentials for various
processes or runtimes that they start, this CL defines two methods
in the testutil/security package:
1) NewCredentials(names): Returns a new Veyron credentials
directory with a principal that has self-signed blessings for 'names'.
2) ForkCredentials(parent, extensions): Returns a new Veyron
credentials directory with a principal that is blessed by 'parent'
under 'extensions'.

Change-Id: I999e0fc0f598574463a66bd0a0023e346f347345