veyron/services/mgmt/device/impl: Fix Claim test and Claim process
With this changes, all the device manager tests use their own temporary
credentials, even when VEYRON_CREDENTIALS is set.

The change also updates TestDeviceManagerClaim such that the
credentials of the unclaimed device manager are independent of the
credentials of the claimant. Without this, the test was rather
meaningless since the claimant already had access to the device.

Since there is no expectation that the unclaimed device's principal and the
claimant will have a common trusted root, the claiming process must also
include trusting the claimant's roots.

Change-Id: If0eff62ec388273d28bac84aedd96e1b90a5aacd