TBR: veyron/services/device/impl,veyron/tools/mgmt: add device uninstall functionality
The uninstall feature of the device manager removes the soft link and the root
dir for the device manager (and pairs with --self_install); the corresponding
dmuninstall script wraps that feature (and pairs with dminstall).

To properly use in the integration test, exposing the pid of the device manager so we can shut it down before running uninstall.

Change-Id: I9c3d079341b35a62da313d82e5bececd1a869906