veyron.io/veyron/veyron2/i18n: Initialize default i18n catalogue from file
This CL introduces an environment variable (VANADIUM_I18N_CATALOGUE) and flag
(-vanadium.i18n_catalogue) to allow the user to specify a comma separated list
of files from which a Vanadium process will read values for the default i18n
catalogue.   These flags are added to veyron.io/veyron/veyron/lib/flags,
alongside -veyron.credentials etc.

The files are read when the runtime is initilized.

It's not clear to me that this is the right place to put the flags, or that the
runtiome initialization is the right place to load the catalogues, but I'm
unsure where else to do it.

The code for writing/parsing catalogue files now always expects the Go format
string "%s %s %q" on each line.  That is, the first two values are never quoted
and the third one (the format string) always is.  This allows escapes in the
format string.

Change-Id: I0ea586b5519e50a9644140ff8c2666eefeaf4bc5