veyron/security/agent: Agent can restart child app upon exit.
Add a flag to instruct the agent to conditionally restart the child command,
based on the value of the command's exit code.  To make this work, defer the
closing of the child unix socket ends in the agent until the agent is ready to
exit. While at it, also close the key mgr socket end (seems like we should).

Add test logic in the integration test to verify this works. Specifically,
ensure both the principal and key manager functions of the agent are usable upon
restart (the latter we verify using vrun).

Change-Id: If1672ebce626ab940464d2ec8ff7955d19d1d761