veyron/services/mgmt/device: device manager Stop/Suspend for handling restart
This change builds on support added to security agent to perform automatic
restarts (https://veyron-review.googlesource.com/8680).  It implements Stop and
Suspend for device manager, with the latter meant to cause the device manager to
exit and be restarted by the agent, whereas the former will cause the agent to
stop and exit.  Plumbing is added to ensure that Stop exits with an appropriate
exit code (presumably, one recognized by the agent to mean 'no more restarts').

impl_test verifies the plumbing for Stop and Suspend.

veyron/tools/mgmt/test.sh: configure the agent to restart the device manager and
verify that Suspend restarts the device manager, and Stop stops it.

Change-Id: Ie3c6294b560fa3c877d825f94901a5da904983a5