veyron/profiles,veyron/runtimes/google/rt/runtimex.go: Create profiles init function for the static profile.
Other profiles can be changed afterwards.

The the new world, a profiles Init func will return a RuntimeX implementation
and a context.

This change also attaches AppCycle, ListenSpec, and Profile to the context.
There is a question as to whether or the Profile object is still needed or
if this func can just replace it.

Change-Id: Iae40cb7a98b8762abf432236fbc21cf1e8784783