veyron/services/security/discharger: Increase validity period of a discharge.
Motivation: Clock skews of more than a minute are quite common. As a
result, a discharge fetched from the service may not be usable at a
server that is more than 1 minute ahead of the discharge service.
(Emprically, our Raspberry Pis were 3 minutes ahead of my laptop)

For now, just punt to 15 minutes!

Change-Id: Ie6f4b5bac3c2d65ea8cca40beb57efa39b254644