veyron/services/mgmt/device/impl: Config override for Install
This change builds on the mechanism introduced in go/vcl/1382 to allow the
application Install'er to override flag settings in the app being installed. The
config is provided as an argument to Install, and persisted with the
installation. We may consider adding the same support to Start as well in the
future.

This change only makes the server-side changes. The corresponding cmd-line tool
changes are left to a future cl.

On the testing side, pack more info in the ping reply to obtain the child app
values for a test flag and test env (the test flag is set and then overriden by
the Install config). While at it, we also get rid of tryInstall in favor of
installAppExpectError (which follows the convention for other rpcs in the impl
test).

Change-Id: Ibeaac48ab82a50ab053fe5d53fc5fb902cfeef71