veyron/services/identity/server: Use the mounted object address instead of the endpoint in the discharger location for third-party caveats created when using BlessUsingAccessToken.
Joint debuggin with nlacasse@

Motivation:
Prior to this change, the identityd running on GCE would end up
setting the DischargerLocation to use an endpoint that is hidden
behind a firewall. What we really want to do is use the mounted
name (that might go through the proxy for example).

HACK ALERT: Because the mounting is asynchronous, for now we
just use the name that the server would like to mount as, not
the name that it successfully mounted as.

Change-Id: Id9948610b24562d66a02a66ae3c0447b0e72bb4e