Add support for handling "unknown" protocols.
There is no longer a default protocol for Endpoints.  Endpoints that do
not have a protocol will have Endpoint.protocol == "".

There is a new method to the stream registry called
"RegisterUnknownProtocol", which is used to register a Dialer and
Listener for endpoints with no protocol.

All the profiles have been updated to use "wsh" for unknown protocols.

The endpoint sorter in sort_endpoints.go will include endpoints
with unknown protocols last.

MultiPart: 1/3

Change-Id: If0771153fc96bffc8dbc79956c8f12f320696343