veyron/lib/unixfd: always set close on exec on the created sockpair
This CL is based on an interpretation of the comment at the top of
https://golang.org/src/syscall/exec_unix.go that suggests that we should always
set close on exec on all files, and exec will then unset that on the ones that
are passed via ExtraFiles. That sounds like the effect we want, but wondering if
there's a "gotcha" that I'm missing. All unit and integration tests pass (on my
linux workstation), but not sure about Mac. If this seems ok, I'll look into
also making all fd's we get through syscall.Dup have close on exec set on them.

Change-Id: Ifcf4b1d0af4a42ef68cfca5ef3f3fa6e3ed0771a