veyron/runtimes/google/ipc: Refactor endpoint sorting.
The primary motivation for this change is to have
filterAndOrderEndpoints work on []naming.MountedServer instead
of []string. This is to ease a future change I plan to make where
naming.MountedServer will also include the BlessingPatterns that the
legitimate server should match and the IPC client will validate servers
it connects to against the pattern in the MountedServer.

Change-Id: I45833157be824542d23e79c39ce147c759140254