tunnel/tunneld: Pass the PATH to the shell
With this change, the PATH env variable of tunneld will be passed to the
shell. For some reason, the shell's rc scripts don't update the PATH
correctly when PATH doesn't exist.

This means that commands like "sudo bash" or "sudo cp" finally work in a
shell that was opened with vsh.

Change-Id: I1a2b938d895004a224c803bb9f4515394d993605